feat(queue):receive messages from service bus

+++ azure/queue/lagerverwaltung/src/validate.ts
@@ -4,12 +4,12 @@ export const productNamePriceValidateRules = [
     body('name','"name" is empty').notEmpty(),
     body('name','Expected type of parameter "name" is String, got another type instead!').isString(),
     body('price','"price is empty').notEmpty(),
-    body('price','Expected type of parameter "price" is numeric, got another type instead!').isNumeric(),
+    body('price','Expected type of parameter "price" is numeric, got another type instead!').isNumeric().toFloat(),
 ];
 
 export const productStockValidate = [
     body('stock','"stock" is empty').notEmpty(),
-    body('stock','Expected type of parameter "stock" is Number (integer positiv), got another type instead!').isInt({min: 0})
+    body('stock','Expected type of parameter "stock" is Number (integer positiv), got another type instead!').isInt({min: 0}).toInt()
 ];
 import { param } from 'express-validator';
 export const idParamValidate = [
